backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
amadeus==8.0.0
redis==5.0.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
dateparser==1.2.0
# Performance optimization dependencies
cachetools==5.3.2
aiohttp==3.9.1
asyncio==3.4.3


